LPD-18833 SF, moves the Upgrade Process to the friendly URL service module as we are modifying its tables without using the service
